rename shared to core
