+++ app/core/CustomError.ts
