+++ app/index.ts
@@ -1,6 +1,6 @@
 import './LoadEnv'; // Must be the first import
 import app from '@server';
-import logger from '@shared/Logger';
+import logger from '@core/Logger';
 
 // Start the server
 const port = Number(process.env.PORT || 3000);
