+++ tsconfig.json
@@ -14,7 +14,7 @@
         "moduleResolution": "node",
         "paths": {
             "@entities/*": ["app/entities/*"],
-            "@shared/*": ["app/shared/*"],
+            "@core/*": ["app/core/*"],
             "@server": ["app/Server"],
             "@config/*": ["app/config/*"]
         },
